refactor: extract tip data and update TipCard styles
