+++ components/dashboard/tip/tip.constants.ts
@@ -1,0 +1,16 @@
+import { Tip } from '@/components/dashboard/tip/tip.types';
+
+export const tips: Tip[] = [
+  {
+    title: 'JavaScript fact',
+    text: '`typeof null` returns "object" — this is a historical bug in JavaScript.',
+  },
+  {
+    title: 'JavaScript fact',
+    text: '`NaN === NaN` is false. Use Number.isNaN() to check it.',
+  },
+  {
+    title: 'JavaScript fact',
+    text: '`[] + []` results in an empty string "" because of type coercion.',
+  },
+];
