+++ components/dashboard/tip/tip.types.ts
@@ -1,0 +1,4 @@
+export type Tip = {
+  title: string;
+  text: string;
+};
